README.md
Ever wanted to wish someone "Happy Birthday" at midnight, but too sleepy and tired to stay awake. Just install the app and fill in the details.
The application as such is self explanatory and in beta stage so it does have bugs. A pretty stand alone app. You can use the code for reference or build upon it.